Makefile.am: fix the order of linker-related parameters of gcc
The -lpcre2-8 parameter was specified at the wrong position so at the end
the build failed with errors like this:

/usr/bin/ld: ./linuxinfo_common.c:254: undefined reference to `pcre2_match_8'

With reordering the parameters in Makefile.am and generating Makefile.in
with automake the problem is now solved.

In ubuntu focal the linuxinfo is not linked to pcre, despite --with-pcre is
explicitly specified in debian/rules as it fails to find the pcre library
when pkg-config is missing. In such case linuxinfo is build without pcre
support.

This issue is already reported to debian but for libpcre, not for linuxinfo:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=962773